requirements.txt
load_dotenv
baidusearch
duckduckgo_search
openai
streamlit